python-cli/requirements.txt
PyInquirer
pyfiglet
transifex-python
